Chapter02/django-myproject/requirements/dev.txt
-r _base.txt 
coverage 
django-debug-toolbar 
selenium
